python_demos/requirements.txt
altgraph==0.17.2
astroid==2.8.4
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
convertdate==2.3.2
Flask==2.0.2
future==0.18.2
hijri-converter==2.2.2
holidays==0.11.3.1
idna==3.3
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
pefile==2021.9.3
platformdirs==2.4.0
pycodestyle==2.8.0
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
pylint==2.11.1
PyMeeus==0.5.11
pyodbc==4.0.32
PySide6==6.2.1
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.26.0
shiboken6==6.2.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
